requirements.txt
ccxt>=4.3.78
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.31.0